Server/requirements.txt
# FastAPI 后端依赖
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# AI 相关
openai==1.3.7

# 文件处理
pdfplumber==0.10.3
python-docx==1.1.0
Pillow==10.1.0

# OCR
pytesseract==0.3.10

# JSON Schema 验证
jsonschema==4.20.0

# 用户认证
bcrypt==4.1.1
PyJWT==2.8.0

# 其他
python-dateutil==2.8.2
